+++ src/api/java/blusunrize/immersiveengineering/api/tool/ToolboxHandler.java
@@ -9,14 +9,20 @@
 package blusunrize.immersiveengineering.api.tool;
 
 import blusunrize.immersiveengineering.api.IETags;
+import com.google.common.collect.Sets;
 import net.minecraft.world.item.DiggerItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ShearsItem;
+import net.minecraftforge.common.ToolAction;
+import net.minecraftforge.common.ToolActions;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 public class ToolboxHandler
 {
@@ -53,9 +59,20 @@ public class ToolboxHandler
 
 	static
 	{
+		Set<ToolAction> toolActions = Sets.newHashSet();
+		toolActions.addAll(ToolActions.DEFAULT_PICKAXE_ACTIONS);
+		toolActions.addAll(ToolActions.DEFAULT_AXE_ACTIONS);
+		toolActions.addAll(ToolActions.DEFAULT_SHOVEL_ACTIONS);
+		toolActions.addAll(ToolActions.DEFAULT_HOE_ACTIONS);
+		toolActions.addAll(ToolActions.DEFAULT_SHEARS_ACTIONS);
+
 		tools.add((s) -> s.is(IETags.toolboxTools));
-		tools.add((s) -> (s.getItem() instanceof DiggerItem));
-		tools.add((s) -> (s.getItem() instanceof ShearsItem));
+		tools.add((s) -> {
+			for(ToolAction action : toolActions)
+				if(s.canPerformAction(action))
+					return true;
+			return false;
+		});
 		foods.add((s) -> (s.getItem().isEdible()));
 		foods.add((s) -> s.is(IETags.toolboxFood));
 		wiring.add((s) -> s.is(IETags.toolboxWiring));